requirements.txt

mysql-connector-python
prettytable
python-dateutil
pandas
requests
pytests
